backend/requirements.txt
fastapi
uvicorn
redis
psycopg2-binary
sqlalchemy
python-jose
passlib[bcrypt]
python-multipart
